memo_creatorクラスを作成・一覧表示機能の処理を修正
・メモの新規作成時のみdb_operatorを直接呼び出していたが、他のオプションと同様に新規作成を担当するクラスを設け、そこから呼び出すようにすることでmemo_appからdb_operatorを認識する必要性をなくした。
・一覧表示機能の実行時、全てのメモを表示した後もプロセスが終了しなくなったため、process.exit(0)を追記して明示的にプロセスを終了させることにした。
