+++ 04.class/memo_creator.js
@@ -1,0 +1,22 @@
+const Memo = require("./memo.js");
+const readline = require("node:readline/promises");
+const DbOperator = require("./db_operator.js");
+const dbOperator = new DbOperator();
+const reader = readline.createInterface({
+  input: process.stdin,
+  output: process.stdout,
+});
+
+module.exports = class MemoCreator {
+  create() {
+    let lines = [];
+    reader.on("line", (line) => {
+      lines.push(line);
+    });
+
+    reader.on("close", () => {
+      const memo = new Memo(lines);
+      dbOperator.create(memo);
+    });
+  }
+}
